src/bot/commands/web_server_commands.py
"""
Discord Commands for Web Server Management
"""
import logging
import asyncio
from discord.ext import commands
from discord import SlashCommandGroup

from ...web.main import app
from ...web.services.redis_service import RedisService

logger = logging.getLogger(__name__)

class WebServerCommands(commands.Cog):
    """Discord commands for web server management"""
    
    def __init__(self, bot):
        self.bot = bot
        self.web_server_process = None
    
    web = SlashCommandGroup("web", "Web server management commands")
    
    @web.command(name="status")
    async def web_status(self, ctx):
        """Check web server status"""
        try:
            # Check if web server is running
            # This would need to be implemented based on your architecture
            await ctx.respond("🌐 Web server status checking...")
            
        except Exception as e:
            await ctx.respond(f"❌ Error checking web server: {e}")
    
    @web.command(name="orders")
    async def web_orders(self, ctx, user_id: str = None):
        """Check recent orders from web interface"""
        try:
            await ctx.defer()
            
            # Get orders from Redis
            # Implementation would connect to your Redis service
            await ctx.followup.send("📋 Order status retrieved")
            
        except Exception as e:
            await ctx.followup.send(f"❌ Error getting orders: {e}")
    
    @web.command(name="signals") 
    async def web_signals(self, ctx):
        """Check recent TradingView signals"""
        try:
            await ctx.defer()
            
            # Get recent signals from Redis
            # Implementation would connect to your Redis service
            await ctx.followup.send("📡 Recent TradingView signals")
            
        except Exception as e:
            await ctx.followup.send(f"❌ Error getting signals: {e}")

async def setup(bot):
    await bot.add_cog(WebServerCommands(bot)) 