requirements.txt
# Discord Trading Bot - Professional Requirements
discord.py>=2.3.0
py-cord>=2.4.0
python-dotenv>=1.0.0
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
asyncio-mqtt>=0.16.0
redis>=4.5.0
SQLAlchemy>=2.0.0
aiohttp>=3.8.0
PyYAML>=6.0
cryptography>=40.0.0
structlog>=23.0.0
psutil>=5.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
alembic>=1.11.0
asyncpg>=0.28.0
pymongo>=4.4.0
celery>=5.3.0

# Security & Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
backtrader>=1.9.76.123
ta-lib>=0.4.25
requests>=2.31.0
schedule>=1.2.0

# FinRL and Deep Reinforcement Learning Dependencies
finrl>=0.3.6
gym>=0.21.0
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.13.0
wandb>=0.15.0
pyfolio>=0.9.2
empyrical>=0.5.5
yfinance>=0.2.18
gymnasium>=0.29.0
sb3-contrib>=2.0.0
optuna>=3.3.0

# Additional dependencies
MetaTrader5>=5.0.0
scipy>=1.10.0
cython>=0.29.0
numba>=0.57.0
prometheus-client>=0.17.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
python-dateutil>=2.8.0
pytz>=2023.3
colorama>=0.4.6